% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/france_departments-data.R
\docType{data}
\name{france_departments}
\alias{france_departments}
\title{France departments population}
\format{
An object of class sf.
}
\usage{
data(france_departments)
}
\description{
Population density for administrative departments of France.
the population density is expressed in number of inhabitants per square kilometer.
}
\keyword{datasets}
