% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/france_regions-data.R
\docType{data}
\name{france_regions}
\alias{france_regions}
\title{France regions population}
\format{
An object of class sf.
}
\usage{
data(france_regions)
}
\description{
Population density for administrative regions of France
}
\keyword{datasets}
