\name{Hotel}
\alias{Hotel}
\docType{data}
\title{Hotel data}
\description{Bertin's hotel data set. Profile of a hotel occupancy for twelve months.}
\usage{data(Hotel)}
\format{
  A data frame with 20 variables for 12 months.
}
\note{For February, the origin proportions sum up to 101\%.}
}
\source{
J. Bertin: \emph{La graphique et le traitement graphique de l'information.} Flammarion: Paris 1977. 
}
\references{A. de Falguerolles; F. Friedrich & G Sawitzki: \emph{A Tribute to J. Bertin's Graphical Data Analysis.}
Published in: W. Bandilla, F. Faulbaum (eds.): SoftStat '97. Advances in Statistical Software 6. 
Lucius & Lucius, Stuttgart 1997
}
\examples{
data(Hotel)
}
\keyword{datasets}

% rm tmp.pdf;  R CMD Rd2dvi --pdf --output=tmp.pdf "/Users/gs/vy/man/Hotel.Rd" ; open tmp.pdf

