% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_points_rst.R
\name{make_points_rst}
\alias{make_points_rst}
\title{Make point valued maps from terra raster object}
\usage{
make_points_rst(rst, n)
}
\arguments{
\item{n}{number of points per side}

\item{rst}{a terra raster object to be transformed as sf point object}
}
\value{
a sf point object
}
\description{
Convert sf polygon object to sf points object
}
