\name{pdfa4w}
\alias{pdfa4w}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{PDF for a4 wide}
\description{Open an a4 PDF port in landscape mode 
}
\usage{
pdfa4w(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{Passed to pdf()}
}
% \details{ If necessary, more details than the description above }
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
  ...
}
%\references{ put references to the literature/web site here  }
\author{G. Sawitzki}
%\note{ further notes 
% Make other sections like Warning with \section{Warning }{....} 
%}
\seealso{pdfa4}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(...) pdf(width=9.952913, height=5.927717, ...)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ device }
